New review for CPSC 110 by ushankab (#555)
> For the most part I enjoyed this course. I feel that it gave me
powerful tools for and ways of thinking about solving problems that I
still use today. In that sense, it does its job a foundational course.
The workload isn’t light. I believe I spent somewhere between 2 and 3
hours a day on 110 on average. The course also jumps in difficulty after
each MT, so it’s important to not get complacent.
>
> Difficulty: 3.5/5
> Quality: 4.5/5
> <cite><a href=''>ushankab</a>, Dec 17 2023, course taken during
2020W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: ushankab
    authorLink: 
    date: 2023-12-17
    review: |
For the most part I enjoyed this course. I feel that it gave me powerful
tools for and ways of thinking about solving problems that I still use
today. In that sense, it does its job a foundational course. The
workload isn’t light. I believe I spent somewhere between 2 and 3 hours
a day on 110 on average. The course also jumps in difficulty after each
MT, so it’s important to not get complacent.
    difficulty: 3.5
    quality: 4.5
    sessionTaken: 2020W1

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker