New review for CPSC 121 by ushankab (#556)
> I’m a big believer in strong fundamentals and foundations and this
course does a decent job of teaching those foundations. Some might not
feel the course is practical, but I’ve used things I learned in this
course on the job to prove the correctness of code I’ve written to
myself. Before taking this course, I wouldn’t have been able to do that.
Also if you like math/logic you will likely enjoy the proof-related
aspects of the course. However, there are some knocks on the course. The
labs and lectures feel largely disjoint, and the course kind of feels
like a hodgepodge of topics than a cohesive whole.
>
> Difficulty: 3.5/5
> Quality: 4/5
> <cite><a href=''>ushankab</a>, Dec 17 2023, course taken during
2021W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: ushankab
    authorLink: 
    date: 2023-12-17
    review: |
I’m a big believer in strong fundamentals and foundations and this
course does a decent job of teaching those foundations. Some might not
feel the course is practical, but I’ve used things I learned in this
course on the job to prove the correctness of code I’ve written to
myself. Before taking this course, I wouldn’t have been able to do that.
Also if you like math/logic you will likely enjoy the proof-related
aspects of the course. However, there are some knocks on the course. The
labs and lectures feel largely disjoint, and the course kind of feels
like a hodgepodge of topics than a cohesive whole.
    difficulty: 3.5
    quality: 4
    sessionTaken: 2021W1

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker