New review for CPSC 340 by Skynet (#563)
> Truly one of the 🤓 courses of all time. Make sure to brush up on your
linear algebra and partial derivatives. Assignments are way harder than
the exams. You will be banging your head over every little detail on the
numpy library. Exams are mostly high level conceptual (recalling facts)
and simple calculations, so if you put effort in the assignments you'll
be prepared for the exams.

Professors are enthusiastic and the course is run well. 
>
> Difficulty: 4/5
> Quality: 5/5
> <cite><a href=''>Skynet</a>, Dec 17 2023, course taken during
2023W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Skynet
    authorLink: 
    date: 2023-12-17
    review: |
Truly one of the 🤓 courses of all time. Make sure to brush up on your
linear algebra and partial derivatives. Assignments are way harder than
the exams. You will be banging your head over every little detail on the
numpy library. Exams are mostly high level conceptual (recalling facts)
and simple calculations, so if you put effort in the assignments you'll
be prepared for the exams.
      
      Professors are enthusiastic and the course is run well. 
    difficulty: 4
    quality: 5
    sessionTaken: 2023W1

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker