New review for CPSC 304 by ushankab (#641)
> One of my favourite CS courses, perhaps even my favourite. If you like
or are interested in databases you will very likely enjoy this course.
As someone who was pretty familiar with databases before I took this
course, I can say the vast majority of the course is very applicable to
the real world. Overall I give top marks to most aspects of the course.
The lectures were well done, the textbook is quite good, there are many
exercises to refine skills, and exams were very reasonable if you paid
attention and did some practice. I do have a few criticisms though. A
few aspects of the project feel somewhat antiquated and suboptimal
(taking a bunch of screenshots of features to put in a Word doc). The
number of errors in the answers to the exercises is higher than should
be the case (not counting the errors corrected in the textbook answer
key). The lecture(s) on datalog could/should be replaced with something
more practical like a survey of NoSQL databases.
>
> Difficulty: 2.5/5
> Quality: 4/5
> <cite><a href=''>ushankab</a>, May 10 2024, course taken during
2023W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: ushankab
    authorLink: 
    date: 2024-05-10
    review: |
One of my favourite CS courses, perhaps even my favourite. If you like
or are interested in databases you will very likely enjoy this course.
As someone who was pretty familiar with databases before I took this
course, I can say the vast majority of the course is very applicable to
the real world. Overall I give top marks to most aspects of the course.
The lectures were well done, the textbook is quite good, there are many
exercises to refine skills, and exams were very reasonable if you paid
attention and did some practice. I do have a few criticisms though. A
few aspects of the project feel somewhat antiquated and suboptimal
(taking a bunch of screenshots of features to put in a Word doc). The
number of errors in the answers to the exercises is higher than should
be the case (not counting the errors corrected in the textbook answer
key). The lecture(s) on datalog could/should be replaced with something
more practical like a survey of NoSQL databases.
    difficulty: 2.5
    quality: 4
    sessionTaken: 2023W2

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker