New review for CPSC 448 by ushankab (#686)
> One of my favourite CPSC courses. Being a self-motivated and
independent learner and having a clear idea of what you want to learn
are important for this course. I used the course to explore a computing
topic that is beyond the scope of the current slate of lecture-based
CPSC courses. I assume this is what many students use this course for,
but perhaps there are other good uses. You do need a (faculty)
supervisor for this course, so it's on you to find someone suitable who
is willing to take you on.
>
> Difficulty: 3/5
> Quality: 5/5
> <cite><a href=''>ushankab</a>, Dec 31 2024, course taken during
2024W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: ushankab
    authorLink: 
    date: 2024-12-31
    review: |
One of my favourite CPSC courses. Being a self-motivated and independent
learner and having a clear idea of what you want to learn are important
for this course. I used the course to explore a computing topic that is
beyond the scope of the current slate of lecture-based CPSC courses. I
assume this is what many students use this course for, but perhaps there
are other good uses. You do need a (faculty) supervisor for this course,
so it's on you to find someone suitable who is willing to take you on.
    difficulty: 3
    quality: 5
    sessionTaken: 2024W1

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker