Added april fools 2025 (#719)
April Fools 2025!

Pretty messy. But simple to add/remove! 

This just uses javascript to replace the HTML in the home page. Can be
controlled with query parameters. The key is `aprilfools`. Disable it
with the value `disable`. Force enable it with `force`.

Still a work in progress! I'll be adding to this over the next week or
so!

I'm embracing BAD code for this. Think of it as method acting.