New review for CPSC 317 by letcodr (#724)
> I think this topic is very fascinating, but it can be a bit dry at
times in lecture. If you don't know much like me, then you will feel a
bit overwhelmed at first because there's so much content and structure
you have to learn. The assignments can be overwhelming for an elective.
>
> Difficulty: 3.5/5
> Quality: 3.5/5
> <cite><a href=''>letcodr</a>, Apr 20 2025, course taken during
2025W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: letcodr
    authorLink: 
    date: 2025-04-20
    review: |
I think this topic is very fascinating, but it can be a bit dry at times
in lecture. If you don't know much like me, then you will feel a bit
overwhelmed at first because there's so much content and structure you
have to learn. The assignments can be overwhelming for an elective.
    difficulty: 3.5
    quality: 3.5
    sessionTaken: 2025W2

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker